Update info53f.md
Filled in. No link to the law.
I am not sure about the scoring. The exception is very limited in practice because it expressly forbids the use of 
- recordings of speeches;
- use of speeches in collections.
I am giving it 2 points because I couldn't find any interpretation on whether or not it covers video recordings and I'm presuming it does since the text excludes sound recordings only.