Update info52d.md
Filled in. Not transposed (IMO).
I have two separate sources stating that 5.2.d. is implemented in Art. L 214-1-2° (but I don't think this is any near ephemeral recordings): 
"Where a phonogram has been published for commercial purposes, the performer and the producer may not oppose: 
(...) 2° Its broadcasting and its simultaneous and integral cable distribution, as well as its reproduction strictly reserved for those purposes, performed by or on behalf of audiovisual communication companies in order to add sound to their own programs broadcast on their own channels as well as on those of audiovisual communication companies who pay fair compensation." 
I need help with the interpretation.