Update info52b.md
Filled in. 

I am a bit at a loss on how to proceed with the German private copying exception, as it is too long and fragmented. I am not sure if I understand correctly some of the provisions - e.g. the requirement for analogue only reproduction in the second part of para 2 (I'm not sure what uses it refers to exactly). 

Also, my source (thus, the link) is not the most up to date. I'm also consulting a book from 2019 but will be grateful for an additional check.