Update info53g.md
Filled in. Not sure about the date. 

Also, my take on the 5(3)(g) implementation differs from most sources incl. the ReCreating table and the book I use ('Copyright in the Information Society (...)' by Brigitte Lindner and Ted Shapiro (eds.)) Most sources cite Section 46 as the implementing provision, but to me it seems that Section 52(2) is more relevant.