Update info53c-1.md
Filled in. Again, left art. 32 paras 1 and 2 out. The so-called 'press clippings' exception is said to apply to 'press reviews', but I don't think it complies with the requirements of the actual 'press review' exception - doesn't have an informative purpose, is not restricted to the use of articles on current events, nor does it provide for an opt-out option for the rightsholder.

Not sure how to interpret the compensation requirement.