Update info53c-1.md
Filled in. I don't have a link to the up-to-date law.
Art.34(c) has one addition that I don't quite understand: "a work so borrowed or the translation thereof may also be used otherwise". 
Does it go beyond the provision of 5.3.c InforSoc? Or does it broaden the scope in terms of beneficiaries?