Update info52e.md
Filled in. No link.
To consult a national expert: I cannot understand what on earth "on the basis of law" would mean in art.34(a). (Hence the 2 points)  
Is it a "work on the basis of law" as in "legal text" or is it "use on the basis of law" as in "lawful use"? Every single one of the resources just recreates the same weird translation.