Update info53c-1.md
I was wandering how to proceed with this one, but ultimately assumed that it was not transposed and added an explication in the 'remarks' section:
The provision of art. 48(1)2 of the Slovenian Copyright Law can be regarded as a 'press review' exception because it covers the 'preparation and reproduction of abstracts of published newspaper and similar articles in the form of press reviews'. However, art. 48(1)2 does not meet two basic requirements of the respective InfoSoc provision of the 1st hypothesis of art.5(3)(c) - (i) the limitation of the thematic scope of the works used to current economic, political or religious topics and (ii) the reservation option in favour of the rightsholder. At the same time, art. 48(1)2 expressly limits the use to the purpose of informing the public, which is a typical characteristic of the 'reporting of current events' exception. This is why it is submitted that the 'press review' exception is not transposed proper in Slovenia and the 'preparation and reproduction of abstracts of published newspaper and similar articles in the form of press reviews' is regarded as part of the 'reporting of current events' exception's implementation.