Update info52e.md
I consider it not transposed. 

The purpose of use of the provision of art 5(2)(e) of the InfoSoc Directive can be recognised in Section 97 - use for the benefit of 'residents or inmates', as well as in Section 98 - use for the benefit of 'clubs, societies and organisations' with charitable objets or are otherwise concerned with the advancement of religion, education or social welfare. However, both provisions provide exclusively for the acts of 'playing' a recording or causing a sound recording, broadcast or cable programme 'to be heard or viewed' and do not mention reproduction at all.